rename package from `check` to `verify` and some tests
