+++ lib/src/test/kotlin/com/lemonappdev/konsist/core/verify/snippet/assert-not-fail.kttxt
