+++ lib/src/test/kotlin/com/lemonappdev/konsist/core/verify/snippet/assert-pass.kttxt
